+++ src/helpers/ImageTransform.php
@@ -56,25 +56,21 @@ class ImageTransform
     // =========================================================================
 
     /**
+     * Transform the $asset for social media sites in $transformName and
+     * optional $siteId
+     *
      * @param int|Asset $asset         the Asset or Asset ID
      * @param string    $transformName the name of the transform to apply
      * @param int|null  $siteId
      *
      * @return string URL to the transformed image
      */
-    public static function socialTransform($asset, string $transformName, int $siteId = null): string
+    public static function socialTransform($asset, string $transformName = '', $siteId = null): string
     {
         $url = '';
-        $config = array_merge(
-            self::$transforms['base'],
-            self::$transforms[$transformName] ?? self::$transforms['base']
-        );
-        $transform = new AssetTransform($config);
-        if (\is_int($asset)) {
-            // Get the asset
-            $asset = Craft::$app->getAssets()->getAssetById($asset, $siteId);
-        }
-        if ($transform && $asset instanceof Asset) {
+        $transform = self::createSocialTransform($transformName);
+        $asset = self::assetFromAssetOrId($asset, $siteId);
+        if (($asset !== null) && ($asset instanceof Asset)) {
             // Generate a transformed image
             $assets = Craft::$app->getAssets();
             $url = $assets->getAssetUrl($asset, $transform, true);
@@ -93,19 +89,12 @@ class ImageTransform
      *
      * @return string width of the transformed image
      */
-    public static function socialTransformWidth($asset, string $transformName, int $siteId = null): string
+    public static function socialTransformWidth($asset, string $transformName = '', $siteId = null): string
     {
         $width = '';
-        $config = array_merge(
-            self::$transforms['base'],
-            self::$transforms[$transformName] ?? self::$transforms['base']
-        );
-        $transform = new AssetTransform($config);
-        if (\is_int($asset)) {
-            // Get the asset
-            $asset = Craft::$app->getAssets()->getAssetById($asset, $siteId);
-        }
-        if ($transform && $asset instanceof Asset) {
+        $transform = self::createSocialTransform($transformName);
+        $asset = self::assetFromAssetOrId($asset, $siteId);
+        if (($asset !== null) && ($asset instanceof Asset)) {
             $width = (string)$asset->getWidth($transform);
             if ($width === null) {
                 $width = '';
@@ -122,19 +111,12 @@ class ImageTransform
      *
      * @return string width of the transformed image
      */
-    public static function socialTransformHeight($asset, string $transformName, int $siteId = null): string
+    public static function socialTransformHeight($asset, string $transformName = '', $siteId = null): string
     {
         $height = '';
-        $config = array_merge(
-            self::$transforms['base'],
-            self::$transforms[$transformName] ?? self::$transforms['base']
-        );
-        $transform = new AssetTransform($config);
-        if (\is_int($asset)) {
-            // Get the asset
-            $asset = Craft::$app->getAssets()->getAssetById($asset, $siteId);
-        }
-        if ($transform && $asset instanceof Asset) {
+        $transform = self::createSocialTransform($transformName);
+        $asset = self::assetFromAssetOrId($asset, $siteId);
+        if (($asset !== null) && ($asset instanceof Asset)) {
             $height = (string)$asset->getHeight($transform);
             if ($height === null) {
                 $height = '';
@@ -152,7 +134,7 @@ class ImageTransform
      *
      * @return array
      */
-    public static function assetElementsFromIds($assetIds, int $siteId = null): array
+    public static function assetElementsFromIds($assetIds, $siteId = null): array
     {
         $elements = Craft::$app->getElements();
         $assets = [];
@@ -169,4 +151,42 @@ class ImageTransform
 
         return $assets;
     }
+
+    // Protected Static Methods
+    // =========================================================================
+
+    /**
+     * Return an asset from either an id or an asset
+     *
+     * @param int|Asset $asset         the Asset or Asset ID
+     * @param int|null  $siteId
+     *
+     * @return Asset|null
+     */
+    protected static function assetFromAssetOrId($asset, $siteId = null)
+    {
+        return \is_int($asset) ? Craft::$app->getAssets()->getAssetById($asset, $siteId) : $asset;
+    }
+
+    /**
+     * Create a transform from the passed in $transformName
+     *
+     * @param string    $transformName the name of the transform to apply
+     *
+     * @return AssetTransform|null
+     */
+    protected static function createSocialTransform(string $transformName = '')
+    {
+        $transform = null;
+        if (!empty($transformName)) {
+            $config = array_merge(
+                self::$transforms['base'],
+                self::$transforms[$transformName] ?? self::$transforms['base']
+            );
+            $transform = new AssetTransform($config);
+        }
+
+        return $transform;
+    }
+
 }